feat: Introduce the `open` command and `--search` flag.
The `open` command is being introduced now as in future there will
likely be other commands (eg. add), and so we don't want people passing
in the search term as the first argument.

`--search` replacing the idea of using the first argument for the search
term so its a little more structured.