fix: Session frequently fails to launch
When using the app, it frequently fails to launch the session.

This is an attempt to fix that by first checking that the session is
ready.

Note: I haven't proven this fixes it as I was unable to recreate the
issue when running from the source. You are committing it to see how it
gets on in the wild.